pyproject.toml
[project]
name = "ibisgraph"
version = "0.1.0"
description = "Pregel-based graph algorithms on top of Ibis project"
readme = "README.md"
requires-python = ">=3.11"


dependencies = [
    "loguru>=0.7",
    "ibis-framework"
]

[dependency-groups]
dev = [
  "ibis-framework[duckdb]",
  "pytest",
  "ruff",
]

[tool.ruff]
line-length = 100
